Breaks out the management of inputs to separate file
